+++ api/src/main/java/com/qaengine/database/QuestionRepository.java
@@ -1,6 +1,6 @@
-package com.kaard.untitled.database;
+package com.qaengine.database;
 
-import com.kaard.untitled.models.Question;
+import com.qaengine.models.Question;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
